pyproject.toml
[tool.poetry]
name = "operator-custom-tg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "operator_custom_tg"}]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "^2.5.3"
pylint = "^2.17.2"
black = "^23.3.0"
python-dotenv = "^1.0.0"
elasticsearch = "^8.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
